Offload event handlers to their own seperate files
