+++ events/warn.js
@@ -1,0 +1,5 @@
+module.exports = {
+  event: require('path').parse(__filename).name,
+  once: false,
+  listener: () => e => (`[WARN] ${e}`)
+};
